requirements.txt
fastapi~=0.116.1
uvicorn
boto3~=1.35.29
python-dotenv~=1.0.1
python-jose
requests~=2.32.4

pydantic~=2.11.7
loguru~=0.7.2
flask~=3.0.3
botocore~=1.35.29
waitress~=3.0.0